+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to [**TNSCM** *(Tenable Nessus CLI Manager)* by LimberDuck][
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2025-09-01
+
+### Added
+
+#### CLI
+
+- New option:
+  - `tnscm --update-check` / `tnscm -u` - will return confirmation if you are using the latest version of TNSCM.
+
+- Requirements update
+  - from:
+    - click>=8.1.8
+    - keyring>=25.5.0
+    - oauthlib>=3.2.2
+    - requests>=2.32.3
+    - pandas>=2.0.3
+  - to:
+    - click>=8.2.1
+    - keyring>=25.6.0
+    - oauthlib>=3.3.1
+    - requests>=2.32.5
+    - pandas>=2.3.2
+
+- tests for python
+  - removed: 3.8, 3.9
+
 ## [0.0.5] - 2025-02-22
 
 ### Changed